Fix up Whiteboard stuff
- Clean up the fetch methods
- Clean up syntax
- Removal of load functions as they add unnecessary complexity
- Small fix when last BU is destroyed in the company page
- Move whiteboard and createWhiteboard pages to seperate components (Later to be used in other pages)
- Pass BURole around in the whiteboard instead of getting it in every component
