+++ Client/my-app/src/routes/companies/+page.svelte
@@ -88,6 +88,9 @@
 
     function handleBUDestroy(BURole) {
         BURoles = BURoles.filter(remove => remove.id !== BURole.id);
+        if (BURole.length === 0) {
+            error = 204;
+        }
     }
 
     onMount(() => {
