+++ src/main/java/com/company/projectManager/whiteboard/whiteboards/controller/WhiteboardController.java
@@ -22,6 +22,8 @@ public class WhiteboardController {
         this.whiteboardService = whiteboardService;
     }
 
+    //TODO: Remove all the endpoints and leave only 1 per method
+
     //Why is this Post? Cuz something something GET requests with bodies is bad (and spring doesn't read it),
     //something something, expose the data in the url (no)
     //I'm either breaking one standard or the other. This one is worse to break but easier to implement. (You might say I'm a total idiot for doing this and I'd agree with you)
@@ -42,6 +44,7 @@ public class WhiteboardController {
 
     @PostMapping(value = {"/company/createWhiteboard", "/company/project/createWhiteboard", "/company/project/team/createWhiteboard"})
     @PreAuthorize("authorityCheck(#whiteboardBUDTO.businessUnitDTO().id(),\"CreateWhiteboard\")")
+    //TODO: This is stupid. There's already a whiteboard inside the BU object
     public ResponseEntity<Object> createWhiteboard(@RequestBody @Valid WhiteboardBusinessUnitDTO whiteboardBUDTO){
         try {
             whiteboardService.createWhiteboard(whiteboardBUDTO.whiteboardDTO(), whiteboardBUDTO.businessUnitDTO());
